Create entries in the audit log outside of the transaction
Since the audit log entries are inserted to the DB in a different goroutine this caused some issues because of how transactions are handled in Gorm.

Also, if an error occurs in the transaction everything will be rolled back, and there whould be no reason to create audit log entries in the first place.